Add files via upload
This project is all about a hospital. Incuding its management panel and user end. This works as user books an appointment with a doctor at their available time , there is no need to visit hospital to book a appointment. This project is fully dymanic.  Asp.net, html, css and javascript is used for creating this  project.